feat: dynamically update type declarations
The customProperties/codeData type declarations are now
dynamically updated based on the json data written in
the customProperties section.
This only works for Grafana v8.2.0-v8.3.X and v9.2.0 and later.
Grafana v8.4.0 updated monaco editor to v0.31.1 which had a bug
which prevents dynamically updating type declarations.
This was fixed in Grafana v9.2.0.