feat: don't wait for declarations to load editor (#203)
Start loading the editor immediately instead of waiting for the
type declarations to be fetched.
The editor can still be used when the declarations are not yet
loaded, but the type declarations are loaded pretty quick,
so it shouldn't be noticeable.

This has kinda the same problems as #202 since this uses
dynamic updating of type declarations.
Straight copied from the commit message:
This only works for Grafana v8.2.0-v8.3.X and v9.2.0 and later.
Grafana v8.4.0 updated monaco editor to v0.31.1 which had a bug
which prevents dynamically updating type declarations.
This was fixed in Grafana v9.2.0.