fix: load declarations with onRender
Load declarations from both onRender and onInit editors.
Only use one as the loader to minimize the number of requests.

This commit also reduces the number of times declarations are loaded.
Before each time the editor was rendered it would load declarations.
Now it only loads them if they are not already loaded.

Closes #201