feat: add adjustable height for the code editor
The height is adjusted with the "resize" css style.
Most browsers support it, but some, like Safari, do not.

The default height is also lowered to 64px mainly to make
scrolling easier.
I also found it a lot easier to get an overview and navigate
through the options with the smaller editor height :D