fix: make grafana validator see declaration paths
The grafana validator seems to have issues seeing files from
require.context, so import the files directly instead.

Error when running the validator before this commit:
The following 1 file(s) differ when comparing source map with source code.
 - ./components/CodeEditor/declarations/ sync \..*\.d\.ts$

The declarations are still moved to a separate folder,
but the import paths are listed explicitly in the source file.